Add interop APIs back to netstandard for PCL/older standard compat
