+++ extensions/interop/interop.cs
@@ -132,11 +132,6 @@ namespace System.Runtime.InteropServices
         CC_STDCALL = 4,
         CC_SYSCALL = 6,
     }
-    [System.AttributeUsageAttribute((System.AttributeTargets)(1028), Inherited=false)]
-    public sealed partial class ComImportAttribute : System.Attribute
-    {
-        public ComImportAttribute() { }
-    }
     [System.ObsoleteAttribute("Use System.Runtime.InteropServices.ComTypes.CONNECTDATA instead. http://go.microsoft.com/fwlink/?linkid=14202", false)]
     [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
     public partial struct CONNECTDATA
@@ -153,11 +148,6 @@ namespace System.Runtime.InteropServices
         DESCKIND_NONE = 0,
         DESCKIND_TYPECOMP = 3,
         DESCKIND_VARDESC = 2,
-    }
-    public sealed partial class DispatchWrapper
-    {
-        public DispatchWrapper(object obj) { }
-        public object WrappedObject { get { throw null; } }
     }
     [System.ObsoleteAttribute("Use System.Runtime.InteropServices.ComTypes.DISPPARAMS instead. http://go.microsoft.com/fwlink/?linkid=14202", false)]
     [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
@@ -182,13 +172,6 @@ namespace System.Runtime.InteropServices
             [System.Runtime.InteropServices.FieldOffsetAttribute(0)]
             public System.Runtime.InteropServices.PARAMDESC paramdesc;
         }
-    }
-    public sealed partial class ErrorWrapper
-    {
-        public ErrorWrapper(System.Exception e) { }
-        public ErrorWrapper(int errorCode) { }
-        public ErrorWrapper(object errorCode) { }
-        public int ErrorCode { get { throw null; } }
     }
     [System.ObsoleteAttribute("Use System.Runtime.InteropServices.ComTypes.EXCEPINFO instead. http://go.microsoft.com/fwlink/?linkid=14202", false)]
     [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
@@ -793,637 +776,5 @@ namespace System.Runtime.InteropServices
         VARFLAG_FRESTRICTED = (short)128,
         VARFLAG_FSOURCE = (short)2,
         VARFLAG_FUIDEFAULT = (short)512,
-    }
-}
-namespace System.Runtime.InteropServices.ComTypes
-{
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct BIND_OPTS
-    {
-        public int cbStruct;
-        public int dwTickCountDeadline;
-        public int grfFlags;
-        public int grfMode;
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Explicit)]
-    public partial struct BINDPTR
-    {
-        [System.Runtime.InteropServices.FieldOffsetAttribute(0)]
-        public System.IntPtr lpfuncdesc;
-        [System.Runtime.InteropServices.FieldOffsetAttribute(0)]
-        public System.IntPtr lptcomp;
-        [System.Runtime.InteropServices.FieldOffsetAttribute(0)]
-        public System.IntPtr lpvardesc;
-    }
-    public enum CALLCONV
-    {
-        CC_CDECL = 1,
-        CC_MACPASCAL = 3,
-        CC_MAX = 9,
-        CC_MPWCDECL = 7,
-        CC_MPWPASCAL = 8,
-        CC_MSCPASCAL = 2,
-        CC_PASCAL = 2,
-        CC_RESERVED = 5,
-        CC_STDCALL = 4,
-        CC_SYSCALL = 6,
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct CONNECTDATA
-    {
-        public int dwCookie;
-        public object pUnk;
-    }
-    public enum DESCKIND
-    {
-        DESCKIND_FUNCDESC = 1,
-        DESCKIND_IMPLICITAPPOBJ = 4,
-        DESCKIND_MAX = 5,
-        DESCKIND_NONE = 0,
-        DESCKIND_TYPECOMP = 3,
-        DESCKIND_VARDESC = 2,
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct DISPPARAMS
-    {
-        public int cArgs;
-        public int cNamedArgs;
-        public System.IntPtr rgdispidNamedArgs;
-        public System.IntPtr rgvarg;
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct ELEMDESC
-    {
-        public System.Runtime.InteropServices.ComTypes.ELEMDESC.DESCUNION desc;
-        public System.Runtime.InteropServices.ComTypes.TYPEDESC tdesc;
-        [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Explicit)]
-        public partial struct DESCUNION
-        {
-            [System.Runtime.InteropServices.FieldOffsetAttribute(0)]
-            public System.Runtime.InteropServices.ComTypes.IDLDESC idldesc;
-            [System.Runtime.InteropServices.FieldOffsetAttribute(0)]
-            public System.Runtime.InteropServices.ComTypes.PARAMDESC paramdesc;
-        }
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct EXCEPINFO
-    {
-        public string bstrDescription;
-        public string bstrHelpFile;
-        public string bstrSource;
-        public int dwHelpContext;
-        public System.IntPtr pfnDeferredFillIn;
-        public System.IntPtr pvReserved;
-        public int scode;
-        public short wCode;
-        public short wReserved;
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct FILETIME
-    {
-        public int dwHighDateTime;
-        public int dwLowDateTime;
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct FUNCDESC
-    {
-        public System.Runtime.InteropServices.ComTypes.CALLCONV callconv;
-        public short cParams;
-        public short cParamsOpt;
-        public short cScodes;
-        public System.Runtime.InteropServices.ComTypes.ELEMDESC elemdescFunc;
-        public System.Runtime.InteropServices.ComTypes.FUNCKIND funckind;
-        public System.Runtime.InteropServices.ComTypes.INVOKEKIND invkind;
-        public System.IntPtr lprgelemdescParam;
-        public System.IntPtr lprgscode;
-        public int memid;
-        public short oVft;
-        public short wFuncFlags;
-    }
-    [System.FlagsAttribute]
-    public enum FUNCFLAGS : short
-    {
-        FUNCFLAG_FBINDABLE = (short)4,
-        FUNCFLAG_FDEFAULTBIND = (short)32,
-        FUNCFLAG_FDEFAULTCOLLELEM = (short)256,
-        FUNCFLAG_FDISPLAYBIND = (short)16,
-        FUNCFLAG_FHIDDEN = (short)64,
-        FUNCFLAG_FIMMEDIATEBIND = (short)4096,
-        FUNCFLAG_FNONBROWSABLE = (short)1024,
-        FUNCFLAG_FREPLACEABLE = (short)2048,
-        FUNCFLAG_FREQUESTEDIT = (short)8,
-        FUNCFLAG_FRESTRICTED = (short)1,
-        FUNCFLAG_FSOURCE = (short)2,
-        FUNCFLAG_FUIDEFAULT = (short)512,
-        FUNCFLAG_FUSESGETLASTERROR = (short)128,
-    }
-    public enum FUNCKIND
-    {
-        FUNC_DISPATCH = 4,
-        FUNC_NONVIRTUAL = 2,
-        FUNC_PUREVIRTUAL = 1,
-        FUNC_STATIC = 3,
-        FUNC_VIRTUAL = 0,
-    }
-    public partial interface IBindCtx
-    {
-        void EnumObjectParam(out System.Runtime.InteropServices.ComTypes.IEnumString ppenum);
-        void GetBindOptions(ref System.Runtime.InteropServices.ComTypes.BIND_OPTS pbindopts);
-        void GetObjectParam(string pszKey, out object ppunk);
-        void GetRunningObjectTable(out System.Runtime.InteropServices.ComTypes.IRunningObjectTable pprot);
-        void RegisterObjectBound(object punk);
-        void RegisterObjectParam(string pszKey, object punk);
-        void ReleaseBoundObjects();
-        void RevokeObjectBound(object punk);
-        int RevokeObjectParam(string pszKey);
-        void SetBindOptions(ref System.Runtime.InteropServices.ComTypes.BIND_OPTS pbindopts);
-    }
-    public partial interface IConnectionPoint
-    {
-        void Advise(object pUnkSink, out int pdwCookie);
-        void EnumConnections(out System.Runtime.InteropServices.ComTypes.IEnumConnections ppEnum);
-        void GetConnectionInterface(out System.Guid pIID);
-        void GetConnectionPointContainer(out System.Runtime.InteropServices.ComTypes.IConnectionPointContainer ppCPC);
-        void Unadvise(int dwCookie);
-    }
-    public partial interface IConnectionPointContainer
-    {
-        void EnumConnectionPoints(out System.Runtime.InteropServices.ComTypes.IEnumConnectionPoints ppEnum);
-        void FindConnectionPoint(ref System.Guid riid, out System.Runtime.InteropServices.ComTypes.IConnectionPoint ppCP);
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct IDLDESC
-    {
-        public System.IntPtr dwReserved;
-        public System.Runtime.InteropServices.ComTypes.IDLFLAG wIDLFlags;
-    }
-    [System.FlagsAttribute]
-    public enum IDLFLAG : short
-    {
-        IDLFLAG_FIN = (short)1,
-        IDLFLAG_FLCID = (short)4,
-        IDLFLAG_FOUT = (short)2,
-        IDLFLAG_FRETVAL = (short)8,
-        IDLFLAG_NONE = (short)0,
-    }
-    public partial interface IEnumConnectionPoints
-    {
-        void Clone(out System.Runtime.InteropServices.ComTypes.IEnumConnectionPoints ppenum);
-        int Next(int celt, System.Runtime.InteropServices.ComTypes.IConnectionPoint[] rgelt, System.IntPtr pceltFetched);
-        void Reset();
-        int Skip(int celt);
-    }
-    public partial interface IEnumConnections
-    {
-        void Clone(out System.Runtime.InteropServices.ComTypes.IEnumConnections ppenum);
-        int Next(int celt, System.Runtime.InteropServices.ComTypes.CONNECTDATA[] rgelt, System.IntPtr pceltFetched);
-        void Reset();
-        int Skip(int celt);
-    }
-    public partial interface IEnumMoniker
-    {
-        void Clone(out System.Runtime.InteropServices.ComTypes.IEnumMoniker ppenum);
-        int Next(int celt, System.Runtime.InteropServices.ComTypes.IMoniker[] rgelt, System.IntPtr pceltFetched);
-        void Reset();
-        int Skip(int celt);
-    }
-    public partial interface IEnumString
-    {
-        void Clone(out System.Runtime.InteropServices.ComTypes.IEnumString ppenum);
-        int Next(int celt, string[] rgelt, System.IntPtr pceltFetched);
-        void Reset();
-        int Skip(int celt);
-    }
-    public partial interface IEnumVARIANT
-    {
-        System.Runtime.InteropServices.ComTypes.IEnumVARIANT Clone();
-        int Next(int celt, object[] rgVar, System.IntPtr pceltFetched);
-        int Reset();
-        int Skip(int celt);
-    }
-    public partial interface IMoniker
-    {
-        void BindToObject(System.Runtime.InteropServices.ComTypes.IBindCtx pbc, System.Runtime.InteropServices.ComTypes.IMoniker pmkToLeft, ref System.Guid riidResult, out object ppvResult);
-        void BindToStorage(System.Runtime.InteropServices.ComTypes.IBindCtx pbc, System.Runtime.InteropServices.ComTypes.IMoniker pmkToLeft, ref System.Guid riid, out object ppvObj);
-        void CommonPrefixWith(System.Runtime.InteropServices.ComTypes.IMoniker pmkOther, out System.Runtime.InteropServices.ComTypes.IMoniker ppmkPrefix);
-        void ComposeWith(System.Runtime.InteropServices.ComTypes.IMoniker pmkRight, bool fOnlyIfNotGeneric, out System.Runtime.InteropServices.ComTypes.IMoniker ppmkComposite);
-        void Enum(bool fForward, out System.Runtime.InteropServices.ComTypes.IEnumMoniker ppenumMoniker);
-        void GetClassID(out System.Guid pClassID);
-        void GetDisplayName(System.Runtime.InteropServices.ComTypes.IBindCtx pbc, System.Runtime.InteropServices.ComTypes.IMoniker pmkToLeft, out string ppszDisplayName);
-        void GetSizeMax(out long pcbSize);
-        void GetTimeOfLastChange(System.Runtime.InteropServices.ComTypes.IBindCtx pbc, System.Runtime.InteropServices.ComTypes.IMoniker pmkToLeft, out System.Runtime.InteropServices.ComTypes.FILETIME pFileTime);
-        void Hash(out int pdwHash);
-        void Inverse(out System.Runtime.InteropServices.ComTypes.IMoniker ppmk);
-        int IsDirty();
-        int IsEqual(System.Runtime.InteropServices.ComTypes.IMoniker pmkOtherMoniker);
-        int IsRunning(System.Runtime.InteropServices.ComTypes.IBindCtx pbc, System.Runtime.InteropServices.ComTypes.IMoniker pmkToLeft, System.Runtime.InteropServices.ComTypes.IMoniker pmkNewlyRunning);
-        int IsSystemMoniker(out int pdwMksys);
-        void Load(System.Runtime.InteropServices.ComTypes.IStream pStm);
-        void ParseDisplayName(System.Runtime.InteropServices.ComTypes.IBindCtx pbc, System.Runtime.InteropServices.ComTypes.IMoniker pmkToLeft, string pszDisplayName, out int pchEaten, out System.Runtime.InteropServices.ComTypes.IMoniker ppmkOut);
-        void Reduce(System.Runtime.InteropServices.ComTypes.IBindCtx pbc, int dwReduceHowFar, ref System.Runtime.InteropServices.ComTypes.IMoniker ppmkToLeft, out System.Runtime.InteropServices.ComTypes.IMoniker ppmkReduced);
-        void RelativePathTo(System.Runtime.InteropServices.ComTypes.IMoniker pmkOther, out System.Runtime.InteropServices.ComTypes.IMoniker ppmkRelPath);
-        void Save(System.Runtime.InteropServices.ComTypes.IStream pStm, bool fClearDirty);
-    }
-    [System.FlagsAttribute]
-    public enum IMPLTYPEFLAGS
-    {
-        IMPLTYPEFLAG_FDEFAULT = 1,
-        IMPLTYPEFLAG_FDEFAULTVTABLE = 8,
-        IMPLTYPEFLAG_FRESTRICTED = 4,
-        IMPLTYPEFLAG_FSOURCE = 2,
-    }
-    [System.FlagsAttribute]
-    public enum INVOKEKIND
-    {
-        INVOKE_FUNC = 1,
-        INVOKE_PROPERTYGET = 2,
-        INVOKE_PROPERTYPUT = 4,
-        INVOKE_PROPERTYPUTREF = 8,
-    }
-    public partial interface IPersistFile
-    {
-        void GetClassID(out System.Guid pClassID);
-        void GetCurFile(out string ppszFileName);
-        int IsDirty();
-        void Load(string pszFileName, int dwMode);
-        void Save(string pszFileName, bool fRemember);
-        void SaveCompleted(string pszFileName);
-    }
-    public partial interface IRunningObjectTable
-    {
-        void EnumRunning(out System.Runtime.InteropServices.ComTypes.IEnumMoniker ppenumMoniker);
-        int GetObject(System.Runtime.InteropServices.ComTypes.IMoniker pmkObjectName, out object ppunkObject);
-        int GetTimeOfLastChange(System.Runtime.InteropServices.ComTypes.IMoniker pmkObjectName, out System.Runtime.InteropServices.ComTypes.FILETIME pfiletime);
-        int IsRunning(System.Runtime.InteropServices.ComTypes.IMoniker pmkObjectName);
-        void NoteChangeTime(int dwRegister, ref System.Runtime.InteropServices.ComTypes.FILETIME pfiletime);
-        int Register(int grfFlags, object punkObject, System.Runtime.InteropServices.ComTypes.IMoniker pmkObjectName);
-        void Revoke(int dwRegister);
-    }
-    public partial interface IStream
-    {
-        void Clone(out System.Runtime.InteropServices.ComTypes.IStream ppstm);
-        void Commit(int grfCommitFlags);
-        void CopyTo(System.Runtime.InteropServices.ComTypes.IStream pstm, long cb, System.IntPtr pcbRead, System.IntPtr pcbWritten);
-        void LockRegion(long libOffset, long cb, int dwLockType);
-        void Read(byte[] pv, int cb, System.IntPtr pcbRead);
-        void Revert();
-        void Seek(long dlibMove, int dwOrigin, System.IntPtr plibNewPosition);
-        void SetSize(long libNewSize);
-        void Stat(out System.Runtime.InteropServices.ComTypes.STATSTG pstatstg, int grfStatFlag);
-        void UnlockRegion(long libOffset, long cb, int dwLockType);
-        void Write(byte[] pv, int cb, System.IntPtr pcbWritten);
-    }
-    public partial interface ITypeComp
-    {
-        void Bind(string szName, int lHashVal, short wFlags, out System.Runtime.InteropServices.ComTypes.ITypeInfo ppTInfo, out System.Runtime.InteropServices.ComTypes.DESCKIND pDescKind, out System.Runtime.InteropServices.ComTypes.BINDPTR pBindPtr);
-        void BindType(string szName, int lHashVal, out System.Runtime.InteropServices.ComTypes.ITypeInfo ppTInfo, out System.Runtime.InteropServices.ComTypes.ITypeComp ppTComp);
-    }
-    public partial interface ITypeInfo
-    {
-        void AddressOfMember(int memid, System.Runtime.InteropServices.ComTypes.INVOKEKIND invKind, out System.IntPtr ppv);
-        void CreateInstance(object pUnkOuter, ref System.Guid riid, out object ppvObj);
-        void GetContainingTypeLib(out System.Runtime.InteropServices.ComTypes.ITypeLib ppTLB, out int pIndex);
-        void GetDllEntry(int memid, System.Runtime.InteropServices.ComTypes.INVOKEKIND invKind, System.IntPtr pBstrDllName, System.IntPtr pBstrName, System.IntPtr pwOrdinal);
-        void GetDocumentation(int index, out string strName, out string strDocString, out int dwHelpContext, out string strHelpFile);
-        void GetFuncDesc(int index, out System.IntPtr ppFuncDesc);
-        void GetIDsOfNames(string[] rgszNames, int cNames, int[] pMemId);
-        void GetImplTypeFlags(int index, out System.Runtime.InteropServices.ComTypes.IMPLTYPEFLAGS pImplTypeFlags);
-        void GetMops(int memid, out string pBstrMops);
-        void GetNames(int memid, string[] rgBstrNames, int cMaxNames, out int pcNames);
-        void GetRefTypeInfo(int hRef, out System.Runtime.InteropServices.ComTypes.ITypeInfo ppTI);
-        void GetRefTypeOfImplType(int index, out int href);
-        void GetTypeAttr(out System.IntPtr ppTypeAttr);
-        void GetTypeComp(out System.Runtime.InteropServices.ComTypes.ITypeComp ppTComp);
-        void GetVarDesc(int index, out System.IntPtr ppVarDesc);
-        void Invoke(object pvInstance, int memid, short wFlags, ref System.Runtime.InteropServices.ComTypes.DISPPARAMS pDispParams, System.IntPtr pVarResult, System.IntPtr pExcepInfo, out int puArgErr);
-        void ReleaseFuncDesc(System.IntPtr pFuncDesc);
-        void ReleaseTypeAttr(System.IntPtr pTypeAttr);
-        void ReleaseVarDesc(System.IntPtr pVarDesc);
-    }
-    public partial interface ITypeInfo2 : System.Runtime.InteropServices.ComTypes.ITypeInfo
-    {
-        new void AddressOfMember(int memid, System.Runtime.InteropServices.ComTypes.INVOKEKIND invKind, out System.IntPtr ppv);
-        new void CreateInstance(object pUnkOuter, ref System.Guid riid, out object ppvObj);
-        void GetAllCustData(System.IntPtr pCustData);
-        void GetAllFuncCustData(int index, System.IntPtr pCustData);
-        void GetAllImplTypeCustData(int index, System.IntPtr pCustData);
-        void GetAllParamCustData(int indexFunc, int indexParam, System.IntPtr pCustData);
-        void GetAllVarCustData(int index, System.IntPtr pCustData);
-        new void GetContainingTypeLib(out System.Runtime.InteropServices.ComTypes.ITypeLib ppTLB, out int pIndex);
-        void GetCustData(ref System.Guid guid, out object pVarVal);
-        new void GetDllEntry(int memid, System.Runtime.InteropServices.ComTypes.INVOKEKIND invKind, System.IntPtr pBstrDllName, System.IntPtr pBstrName, System.IntPtr pwOrdinal);
-        new void GetDocumentation(int index, out string strName, out string strDocString, out int dwHelpContext, out string strHelpFile);
-        [System.Runtime.InteropServices.LCIDConversionAttribute(1)]
-        void GetDocumentation2(int memid, out string pbstrHelpString, out int pdwHelpStringContext, out string pbstrHelpStringDll);
-        void GetFuncCustData(int index, ref System.Guid guid, out object pVarVal);
-        new void GetFuncDesc(int index, out System.IntPtr ppFuncDesc);
-        void GetFuncIndexOfMemId(int memid, System.Runtime.InteropServices.ComTypes.INVOKEKIND invKind, out int pFuncIndex);
-        new void GetIDsOfNames(string[] rgszNames, int cNames, int[] pMemId);
-        void GetImplTypeCustData(int index, ref System.Guid guid, out object pVarVal);
-        new void GetImplTypeFlags(int index, out System.Runtime.InteropServices.ComTypes.IMPLTYPEFLAGS pImplTypeFlags);
-        new void GetMops(int memid, out string pBstrMops);
-        new void GetNames(int memid, string[] rgBstrNames, int cMaxNames, out int pcNames);
-        void GetParamCustData(int indexFunc, int indexParam, ref System.Guid guid, out object pVarVal);
-        new void GetRefTypeInfo(int hRef, out System.Runtime.InteropServices.ComTypes.ITypeInfo ppTI);
-        new void GetRefTypeOfImplType(int index, out int href);
-        new void GetTypeAttr(out System.IntPtr ppTypeAttr);
-        new void GetTypeComp(out System.Runtime.InteropServices.ComTypes.ITypeComp ppTComp);
-        void GetTypeFlags(out int pTypeFlags);
-        void GetTypeKind(out System.Runtime.InteropServices.ComTypes.TYPEKIND pTypeKind);
-        void GetVarCustData(int index, ref System.Guid guid, out object pVarVal);
-        new void GetVarDesc(int index, out System.IntPtr ppVarDesc);
-        void GetVarIndexOfMemId(int memid, out int pVarIndex);
-        new void Invoke(object pvInstance, int memid, short wFlags, ref System.Runtime.InteropServices.ComTypes.DISPPARAMS pDispParams, System.IntPtr pVarResult, System.IntPtr pExcepInfo, out int puArgErr);
-        new void ReleaseFuncDesc(System.IntPtr pFuncDesc);
-        new void ReleaseTypeAttr(System.IntPtr pTypeAttr);
-        new void ReleaseVarDesc(System.IntPtr pVarDesc);
-    }
-    public partial interface ITypeLib
-    {
-        void FindName(string szNameBuf, int lHashVal, System.Runtime.InteropServices.ComTypes.ITypeInfo[] ppTInfo, int[] rgMemId, ref short pcFound);
-        void GetDocumentation(int index, out string strName, out string strDocString, out int dwHelpContext, out string strHelpFile);
-        void GetLibAttr(out System.IntPtr ppTLibAttr);
-        void GetTypeComp(out System.Runtime.InteropServices.ComTypes.ITypeComp ppTComp);
-        void GetTypeInfo(int index, out System.Runtime.InteropServices.ComTypes.ITypeInfo ppTI);
-        int GetTypeInfoCount();
-        void GetTypeInfoOfGuid(ref System.Guid guid, out System.Runtime.InteropServices.ComTypes.ITypeInfo ppTInfo);
-        void GetTypeInfoType(int index, out System.Runtime.InteropServices.ComTypes.TYPEKIND pTKind);
-        bool IsName(string szNameBuf, int lHashVal);
-        void ReleaseTLibAttr(System.IntPtr pTLibAttr);
-    }
-    public partial interface ITypeLib2 : System.Runtime.InteropServices.ComTypes.ITypeLib
-    {
-        new void FindName(string szNameBuf, int lHashVal, System.Runtime.InteropServices.ComTypes.ITypeInfo[] ppTInfo, int[] rgMemId, ref short pcFound);
-        void GetAllCustData(System.IntPtr pCustData);
-        void GetCustData(ref System.Guid guid, out object pVarVal);
-        new void GetDocumentation(int index, out string strName, out string strDocString, out int dwHelpContext, out string strHelpFile);
-        [System.Runtime.InteropServices.LCIDConversionAttribute(1)]
-        void GetDocumentation2(int index, out string pbstrHelpString, out int pdwHelpStringContext, out string pbstrHelpStringDll);
-        new void GetLibAttr(out System.IntPtr ppTLibAttr);
-        void GetLibStatistics(System.IntPtr pcUniqueNames, out int pcchUniqueNames);
-        new void GetTypeComp(out System.Runtime.InteropServices.ComTypes.ITypeComp ppTComp);
-        new void GetTypeInfo(int index, out System.Runtime.InteropServices.ComTypes.ITypeInfo ppTI);
-        new int GetTypeInfoCount();
-        new void GetTypeInfoOfGuid(ref System.Guid guid, out System.Runtime.InteropServices.ComTypes.ITypeInfo ppTInfo);
-        new void GetTypeInfoType(int index, out System.Runtime.InteropServices.ComTypes.TYPEKIND pTKind);
-        new bool IsName(string szNameBuf, int lHashVal);
-        new void ReleaseTLibAttr(System.IntPtr pTLibAttr);
-    }
-    [System.FlagsAttribute]
-    public enum LIBFLAGS : short
-    {
-        LIBFLAG_FCONTROL = (short)2,
-        LIBFLAG_FHASDISKIMAGE = (short)8,
-        LIBFLAG_FHIDDEN = (short)4,
-        LIBFLAG_FRESTRICTED = (short)1,
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct PARAMDESC
-    {
-        public System.IntPtr lpVarValue;
-        public System.Runtime.InteropServices.ComTypes.PARAMFLAG wParamFlags;
-    }
-    [System.FlagsAttribute]
-    public enum PARAMFLAG : short
-    {
-        PARAMFLAG_FHASCUSTDATA = (short)64,
-        PARAMFLAG_FHASDEFAULT = (short)32,
-        PARAMFLAG_FIN = (short)1,
-        PARAMFLAG_FLCID = (short)4,
-        PARAMFLAG_FOPT = (short)16,
-        PARAMFLAG_FOUT = (short)2,
-        PARAMFLAG_FRETVAL = (short)8,
-        PARAMFLAG_NONE = (short)0,
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct STATSTG
-    {
-        public System.Runtime.InteropServices.ComTypes.FILETIME atime;
-        public long cbSize;
-        public System.Guid clsid;
-        public System.Runtime.InteropServices.ComTypes.FILETIME ctime;
-        public int grfLocksSupported;
-        public int grfMode;
-        public int grfStateBits;
-        public System.Runtime.InteropServices.ComTypes.FILETIME mtime;
-        public string pwcsName;
-        public int reserved;
-        public int type;
-    }
-    public enum SYSKIND
-    {
-        SYS_MAC = 2,
-        SYS_WIN16 = 0,
-        SYS_WIN32 = 1,
-        SYS_WIN64 = 3,
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct TYPEATTR
-    {
-        public short cbAlignment;
-        public int cbSizeInstance;
-        public short cbSizeVft;
-        public short cFuncs;
-        public short cImplTypes;
-        public short cVars;
-        public int dwReserved;
-        public System.Guid guid;
-        public System.Runtime.InteropServices.ComTypes.IDLDESC idldescType;
-        public int lcid;
-        public System.IntPtr lpstrSchema;
-        public const int MEMBER_ID_NIL = -1;
-        public int memidConstructor;
-        public int memidDestructor;
-        public System.Runtime.InteropServices.ComTypes.TYPEDESC tdescAlias;
-        public System.Runtime.InteropServices.ComTypes.TYPEKIND typekind;
-        public short wMajorVerNum;
-        public short wMinorVerNum;
-        public System.Runtime.InteropServices.ComTypes.TYPEFLAGS wTypeFlags;
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct TYPEDESC
-    {
-        public System.IntPtr lpValue;
-        public short vt;
-    }
-    [System.FlagsAttribute]
-    public enum TYPEFLAGS : short
-    {
-        TYPEFLAG_FAGGREGATABLE = (short)1024,
-        TYPEFLAG_FAPPOBJECT = (short)1,
-        TYPEFLAG_FCANCREATE = (short)2,
-        TYPEFLAG_FCONTROL = (short)32,
-        TYPEFLAG_FDISPATCHABLE = (short)4096,
-        TYPEFLAG_FDUAL = (short)64,
-        TYPEFLAG_FHIDDEN = (short)16,
-        TYPEFLAG_FLICENSED = (short)4,
-        TYPEFLAG_FNONEXTENSIBLE = (short)128,
-        TYPEFLAG_FOLEAUTOMATION = (short)256,
-        TYPEFLAG_FPREDECLID = (short)8,
-        TYPEFLAG_FPROXY = (short)16384,
-        TYPEFLAG_FREPLACEABLE = (short)2048,
-        TYPEFLAG_FRESTRICTED = (short)512,
-        TYPEFLAG_FREVERSEBIND = (short)8192,
-    }
-    public enum TYPEKIND
-    {
-        TKIND_ALIAS = 6,
-        TKIND_COCLASS = 5,
-        TKIND_DISPATCH = 4,
-        TKIND_ENUM = 0,
-        TKIND_INTERFACE = 3,
-        TKIND_MAX = 8,
-        TKIND_MODULE = 2,
-        TKIND_RECORD = 1,
-        TKIND_UNION = 7,
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct TYPELIBATTR
-    {
-        public System.Guid guid;
-        public int lcid;
-        public System.Runtime.InteropServices.ComTypes.SYSKIND syskind;
-        public System.Runtime.InteropServices.ComTypes.LIBFLAGS wLibFlags;
-        public short wMajorVerNum;
-        public short wMinorVerNum;
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct VARDESC
-    {
-        public System.Runtime.InteropServices.ComTypes.VARDESC.DESCUNION desc;
-        public System.Runtime.InteropServices.ComTypes.ELEMDESC elemdescVar;
-        public string lpstrSchema;
-        public int memid;
-        public System.Runtime.InteropServices.ComTypes.VARKIND varkind;
-        public short wVarFlags;
-        [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Explicit)]
-        public partial struct DESCUNION
-        {
-            [System.Runtime.InteropServices.FieldOffsetAttribute(0)]
-            public System.IntPtr lpvarValue;
-            [System.Runtime.InteropServices.FieldOffsetAttribute(0)]
-            public int oInst;
-        }
-    }
-    [System.FlagsAttribute]
-    public enum VARFLAGS : short
-    {
-        VARFLAG_FBINDABLE = (short)4,
-        VARFLAG_FDEFAULTBIND = (short)32,
-        VARFLAG_FDEFAULTCOLLELEM = (short)256,
-        VARFLAG_FDISPLAYBIND = (short)16,
-        VARFLAG_FHIDDEN = (short)64,
-        VARFLAG_FIMMEDIATEBIND = (short)4096,
-        VARFLAG_FNONBROWSABLE = (short)1024,
-        VARFLAG_FREADONLY = (short)1,
-        VARFLAG_FREPLACEABLE = (short)2048,
-        VARFLAG_FREQUESTEDIT = (short)8,
-        VARFLAG_FRESTRICTED = (short)128,
-        VARFLAG_FSOURCE = (short)2,
-        VARFLAG_FUIDEFAULT = (short)512,
-    }
-    public enum VARKIND
-    {
-        VAR_CONST = 2,
-        VAR_DISPATCH = 3,
-        VAR_PERINSTANCE = 0,
-        VAR_STATIC = 1,
-    }
-}
-namespace System.Runtime.InteropServices.ComTypes
-{
-    [System.FlagsAttribute]
-    public enum ADVF
-    {
-        ADVF_DATAONSTOP = 64,
-        ADVF_NODATA = 1,
-        ADVF_ONLYONCE = 4,
-        ADVF_PRIMEFIRST = 2,
-        ADVFCACHE_FORCEBUILTIN = 16,
-        ADVFCACHE_NOHANDLER = 8,
-        ADVFCACHE_ONSAVE = 32,
-    }
-    public enum DATADIR
-    {
-        DATADIR_GET = 1,
-        DATADIR_SET = 2,
-    }
-    [System.FlagsAttribute]
-    public enum DVASPECT
-    {
-        DVASPECT_CONTENT = 1,
-        DVASPECT_DOCPRINT = 8,
-        DVASPECT_ICON = 4,
-        DVASPECT_THUMBNAIL = 2,
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct FORMATETC
-    {
-        public short cfFormat;
-        public System.Runtime.InteropServices.ComTypes.DVASPECT dwAspect;
-        public int lindex;
-        public System.IntPtr ptd;
-        public System.Runtime.InteropServices.ComTypes.TYMED tymed;
-    }
-    public partial interface IAdviseSink
-    {
-        void OnClose();
-        void OnDataChange(ref System.Runtime.InteropServices.ComTypes.FORMATETC format, ref System.Runtime.InteropServices.ComTypes.STGMEDIUM stgmedium);
-        void OnRename(System.Runtime.InteropServices.ComTypes.IMoniker moniker);
-        void OnSave();
-        void OnViewChange(int aspect, int index);
-    }
-    public partial interface IDataObject
-    {
-        int DAdvise(ref System.Runtime.InteropServices.ComTypes.FORMATETC pFormatetc, System.Runtime.InteropServices.ComTypes.ADVF advf, System.Runtime.InteropServices.ComTypes.IAdviseSink adviseSink, out int connection);
-        void DUnadvise(int connection);
-        int EnumDAdvise(out System.Runtime.InteropServices.ComTypes.IEnumSTATDATA enumAdvise);
-        System.Runtime.InteropServices.ComTypes.IEnumFORMATETC EnumFormatEtc(System.Runtime.InteropServices.ComTypes.DATADIR direction);
-        int GetCanonicalFormatEtc(ref System.Runtime.InteropServices.ComTypes.FORMATETC formatIn, out System.Runtime.InteropServices.ComTypes.FORMATETC formatOut);
-        void GetData(ref System.Runtime.InteropServices.ComTypes.FORMATETC format, out System.Runtime.InteropServices.ComTypes.STGMEDIUM medium);
-        void GetDataHere(ref System.Runtime.InteropServices.ComTypes.FORMATETC format, ref System.Runtime.InteropServices.ComTypes.STGMEDIUM medium);
-        int QueryGetData(ref System.Runtime.InteropServices.ComTypes.FORMATETC format);
-        void SetData(ref System.Runtime.InteropServices.ComTypes.FORMATETC formatIn, ref System.Runtime.InteropServices.ComTypes.STGMEDIUM medium, bool release);
-    }
-    public partial interface IEnumFORMATETC
-    {
-        void Clone(out System.Runtime.InteropServices.ComTypes.IEnumFORMATETC newEnum);
-        int Next(int celt, System.Runtime.InteropServices.ComTypes.FORMATETC[] rgelt, int[] pceltFetched);
-        int Reset();
-        int Skip(int celt);
-    }
-    public partial interface IEnumSTATDATA
-    {
-        void Clone(out System.Runtime.InteropServices.ComTypes.IEnumSTATDATA newEnum);
-        int Next(int celt, System.Runtime.InteropServices.ComTypes.STATDATA[] rgelt, int[] pceltFetched);
-        int Reset();
-        int Skip(int celt);
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct STATDATA
-    {
-        public System.Runtime.InteropServices.ComTypes.ADVF advf;
-        public System.Runtime.InteropServices.ComTypes.IAdviseSink advSink;
-        public int connection;
-        public System.Runtime.InteropServices.ComTypes.FORMATETC formatetc;
-    }
-    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
-    public partial struct STGMEDIUM
-    {
-        public object pUnkForRelease;
-        public System.Runtime.InteropServices.ComTypes.TYMED tymed;
-        public System.IntPtr unionmember;
-    }
-    [System.FlagsAttribute]
-    public enum TYMED
-    {
-        TYMED_ENHMF = 64,
-        TYMED_FILE = 2,
-        TYMED_GDI = 16,
-        TYMED_HGLOBAL = 1,
-        TYMED_ISTORAGE = 8,
-        TYMED_ISTREAM = 4,
-        TYMED_MFPICT = 32,
-        TYMED_NULL = 0,
     }
 }
